+++ CHANGELOG.md
@@ -17,9 +17,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- reduce Filter classes to a single list field of `FilterField` items
-
 ### Security
+
+## [0.49.1] - 2025-01-29
+
+### Fixed
+
+- reduce Filter classes to a single list field of `FilterField` items
 
 ## [0.49.0] - 2025-01-29